requirements.txt
asgiref
crispy-bootstrap5
distlib
dj-database-url
Django
django-cors-headers
django-crispy-forms
django-environ
django-filter
django-humanize
django-multiselectfield
djangorestframework
filelock
humanize
Markdown
pillow
platformdirs
psycopg2-binary
sqlparse
typing_extensions
tzdata
virtualenv
